+++ pukr/__init__.py
@@ -1,4 +1,5 @@
 from .config import get_logger
+from .custom_handlers import InterceptHandler
 
-__version__ = '0.0.2'
-__update_date__ = '2022-05-01_19:59:20'
+__version__ = '0.0.3'
+__update_date__ = '2022-06-25_16:47:36'